spike(skydive): 날개 자세를 소수 틱으로도 물을 수 있게 + 계측
두 가지다.

(1) 떨림. 시뮬은 50Hz인데 화면은 60Hz 이상이라, 틱 자세만 쓰면 여섯 프레임 중
하나가 제자리라 계단처럼 떤다(추격자 벽이 같은 이유로 0.2m씩 계단으로 왔다).
각도 식이 소수 틱도 받게 해서 뷰가 프레임 사이 자세를 물을 수 있게 했다.
식을 하나로 두는 이유는 둘이 갈라지면 보이는 자세와 맞는 자세가 어긋나기 때문이다.
판정은 안 바뀐다 — 다음 틱에 시뮬이 정수 틱 자세로 다시 잡고 그 자세에서 질의한다.

(2) [진단용 임시] 날개 원 안에 있는 동안 매 틱 정황을 남긴다. 클·서가 같은
형식으로 찍으므로 같은 tick 줄끼리 대조하면 어디서 갈리는지가 드러난다 —
날개 각도가 다른지, 밀어낸 벡터가 다른지, 이동 sweep이 지운 속도가 다른지.

실플레이에서 리컨 스파이크 6건이 전부 팬 원 안(중심 1.6~2.9m)에서 났고,
위치는 3~5cm인데 속도는 방향이 다를 만큼 갈렸다. 밀어내기는 속도를 안
건드리므로 이동 sweep 쪽이 의심되지만, 재서 가른다.

+++ Runtime/Scripts/Game/BladeGeometry.cs
@@ -14,7 +14,9 @@ namespace LOP
         /// 그때 `속도 × 틱`을 float로 그냥 두면 자릿수가 커져 정밀도가 뭉개진다. 접는 계산만
         /// double로 하고 결과는 작은 값이라 float로 안전하다.
         /// </summary>
-        public static float AngleDegreesAt(float startDegrees, float speedDegreesPerTick, long tick)
+        /// <param name="tick">판정은 정수 틱, 그림은 소수 틱(프레임 사이)을 넣는다.
+        /// <b>식을 하나로 두는 이유</b>: 둘이 갈라지면 보이는 자세와 맞는 자세가 어긋난다.</param>
+        public static float AngleDegreesAt(float startDegrees, float speedDegreesPerTick, double tick)
         {
             double raw = startDegrees + (double)speedDegreesPerTick * tick;
             double folded = raw - System.Math.Floor(raw / 360.0) * 360.0;
